+++ config.yaml
@@ -1,5 +1,7 @@
 DB_PASS: "..." #вставьте свой пароль от базы данных
 GROUP_TOKEN: "vk1.a.BwzhpjkWPeWTC0mq-8QfU5jrldMLXpZGGT1BVqyLQsnmtqbuexlYlG9rUrhZQy1gZb7abI6eQ8i4zvlG4xZwJ0d8o1SjS7tdK6YredgoGVL9uOq8ZJW0yNyROMje_zynfG7FY-KahOhnC-KrSygld_blZ4CXCyJdSMtlsdGePE7WoVIzx9c-TmKSfhmcWaHZAZTc0ShcKWsRzljZj7FMqw"
-ACCESS_TOKEN: "..."#вставьте свой персональный пользовательский токен вк
+ACCESS_TOKEN: "..."
+
+#вставьте свой персональный пользовательский токен вк
